Dodano obsługę błędów

+++ Services/UzytkownikService.cs
@@ -12,13 +12,25 @@ namespace krzysztofb.Services
         IDatabaseCreate<UzytkownikDTO>
     {
         private readonly WnioskiContext _context;
+
         public UzytkownikService(WnioskiContext context)
         {
             _context = context;
         }
         public UzytkownikDTO Create(UzytkownikDTO obj)
         {
-
+            if (_context.Uzytkownik.Find(obj.IdPrzelozonego) == null)
+            {
+                throw new BadHttpRequestException("User with given idPrzelozony not found");
+            }
+            else if (_context.Role.Find(obj.Role) == null)
+            {
+                throw new BadHttpRequestException("Role not found");
+            }
+            else if (_context.Uzytkownik.FirstOrDefault(x => x.Email == obj.Email) != null)
+            {
+                throw new DbUpdateException("Email is not unique");
+            }
             _context.Uzytkownik.Add(ModelConverter.ConvertToModel(obj));
             _context.SaveChanges();
             return obj;
@@ -27,7 +39,10 @@ namespace krzysztofb.Services
         public UzytkownikDTO Delete(int id)
         {
             var user = _context.Uzytkownik.Find(id);
-
+            if (user == null)
+            {
+                throw new BadHttpRequestException("User not found");
+            }
             _context.Uzytkownik.Remove(user);
             return ModelConverter.ConvertToDTO(user);
         }
@@ -43,6 +58,10 @@ namespace krzysztofb.Services
         }
         public UzytkownikDTO Read(int id)
         {
+            if (_context.Uzytkownik.Find(id) == null)
+            {
+                throw new NullReferenceException("User not found");
+            }
             var user = _context.Uzytkownik
                 .Include(x => x.IdPrzelozonegoNavigation)
                 .Include(x => x.RoleNavigation)
@@ -51,6 +70,22 @@ namespace krzysztofb.Services
         }
         public UzytkownikDTO Update(int id, UzytkownikDTO obj)
         {
+            if (_context.Uzytkownik.Find(id) == null)
+            {
+                throw new BadHttpRequestException("User not found");
+            }
+            else if (_context.Uzytkownik.Find(obj.IdPrzelozonego) == null)
+            {
+                throw new BadHttpRequestException("User with given idPrzelozony not found");
+            }
+            else if (_context.Role.Find(obj.Role) == null)
+            {
+                throw new BadHttpRequestException("Role not found");
+            }
+            else if (_context.Uzytkownik.FirstOrDefault(x => x.Email == obj.Email) != null)
+            {
+                throw new DbUpdateException("Email is not unique");
+            }
             var user = _context.Uzytkownik.Find(id);
 
             foreach (var propetryEntry in _context.Entry(user).Properties)
